Added fixture to conftest
