+++ gladier_tools/tests/posix/conftest.py
@@ -1,3 +1,12 @@
 import pytest
 import tarfile
-from test_tar import mock_tar
+from unittest.mock import Mock
+
+@pytest.fixture
+def mock_tar(monkeypatch):
+    mock_tf = Mock()
+    mock_open = Mock()
+    mock_open.return_value.__enter__ = Mock(return_value=mock_tf)
+    mock_open.return_value.__exit__ = Mock(return_value=None)
+    monkeypatch.setattr(tarfile, 'open', mock_open)
+    return mock_open, mock_tf
